moved calculate score function
